More logical places for the logic, fix the tests
