+++ hq/app/logic/ReportDisplay.scala
@@ -1,12 +1,9 @@
 package logic
 
-import com.amazonaws.services.cloudformation.model.{DescribeStacksResult, Output, Stack}
 import logic.DateUtils.dayDiff
 import model._
 import org.joda.time.{DateTime, DateTimeZone, Days}
 import utils.attempt.FailedAttempt
-
-import scala.collection.JavaConverters._
 
 
 object ReportDisplay {
@@ -53,30 +50,9 @@ object ReportDisplay {
     else Green
   }
 
-  private[logic] def getUserStacks(stacks: List[Stack]): List[Stack] = {
-    def outputsUser(output: Output): Boolean = {
-      output.getOutputValue.contains(":user")
-    }
-    for {
-      stack <- stacks
-      output <- stack.getOutputs.asScala.toList
-      if outputsUser(output)
-    } yield stack
-  }
+  def toCredentialReportDisplay(report: IAMCredentialsReport): CredentialReportDisplay = {
 
-  private[logic] def enrichUserDetails(credentials: Seq[IAMCredential], stacks: List[Stack]): Seq[IAMCredential] = {
-    credentials.map { cred =>
-      val enrichedCred = for {
-        output <- stacks.find(_.getOutputs.asScala.toList.exists(_.getOutputValue.contains(cred.arn)))
-      } yield cred.copy(stackId = Some(output.getStackId), stackName = Some(output.getStackName))
-      enrichedCred.getOrElse(cred)
-    }
-  }
-
-  def toCredentialReportDisplay(report: IAMCredentialsReport, stacks: DescribeStacksResult): CredentialReportDisplay = {
-    val userStacks = getUserStacks(stacks.getStacks.asScala.toList)
-
-    enrichUserDetails(report.entries, userStacks).filterNot(_.rootUser).foldLeft(CredentialReportDisplay(report.generatedAt)) { (report, cred) =>
+    report.entries.filterNot(_.rootUser).foldLeft(CredentialReportDisplay(report.generatedAt)) { (report, cred) =>
       val machineUser =
         if (!cred.passwordEnabled.getOrElse(false)) {
           Some(MachineUser(
