+++ hq/app/views/iam/iamCredReportBody.scala.html
@@ -25,7 +25,7 @@
                     <td>
                         <i class="material-icons left tooltipped" data-tooltip="human user" data-position="bottom" data-delay="50">person</i>
                         @if(cred.stackId && cred.stackName) {
-                            <i class="material-icons left tooltipped" data-tooltip="cloudformed" data-position="bottom" data-delay="50">cloud</i>
+                            <i class="material-icons left tooltipped" data-tooltip="cloudformed" data-position="bottom" data-delay="50">cloud_done</i>
                         } else {
                             <i class="material-icons left tooltipped" data-tooltip="not cloudformed" data-position="bottom" data-delay="50">cloud_off</i>
                         }
@@ -61,9 +61,9 @@
                     <td>
                         <i class="material-icons left tooltipped" data-tooltip="machine user" data-position="bottom" data-delay="50">computer</i>
                         @if(cred.stackId && cred.stackName) {
-                        <i class="material-icons left tooltipped" data-tooltip="cloudformed" data-position="bottom" data-delay="50">cloud</i>
+                        <i class="material-icons left tooltipped" data-tooltip="cloudformed" data-position="bottom" data-delay="50">cloud_done</i>
                         } else {
-                        <i class="material-icons left tooltipped" data-tooltip="cloudformed" data-position="bottom" data-delay="50">cloud_off</i>
+                        <i class="material-icons left tooltipped" data-tooltip="not cloudformed" data-position="bottom" data-delay="50">cloud_off</i>
                         }
                     </td>
                     <td>
