+++ hq/test/aws/iam/CredentialsReportTest.scala
@@ -91,7 +91,6 @@ class CredentialsReportTest extends FreeSpec with Matchers with OptionValues wit
         )
       }
 
-
       "fails when parsing empty CSV" in {
         val testReport = ""
         CredentialsReport.tryParsingReport(testReport).isFailedAttempt() shouldBe true
@@ -108,9 +107,7 @@ class CredentialsReportTest extends FreeSpec with Matchers with OptionValues wit
         userWithCreds should have (
           'passwordLastUsed (None)
         )
-
       }
-
     }
   }
 }
